pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Michael Roberts - CSAcademyIO",email = "m.roberts94@hotmail.co.uk"}
]
readme = "README.md"
requires-python = ">=3.10, < 4.0"
dependencies = [
    "pydantic (>=2.11.9,<3.0.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "aioredis (>=2.0.1,<3.0.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "uuid (>=1.30,<2.0)",
    "fastapi (>=0.117.1,<0.118.0)",
    "aiokafka (>=0.12.0,<0.13.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "grpcio (>=1.75.1,<2.0.0)",
    "grpcio-tools (>=1.75.1,<2.0.0)",
    "grpcio-reflection (>=1.75.1,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[brcrypt] (>=1.7.4,<2.0.0)",
    "geoalchemy2 (>=0.18.0,<0.19.0)",
    "dynaconf (>=3.2.11,<4.0.0)",
    "phonenumbers (>=9.0.15,<10.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "jwt (>=1.4.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
